+++ src/Farmer/Arm/Web.fs
@@ -198,7 +198,8 @@ type Site =
       Metadata : List<string * string>
       AutoSwapSlotName: string option
       ZipDeployPath : (string * ZipDeploy.ZipDeployTarget * ZipDeploy.ZipDeploySlot) option
-      HealthCheckPath : string option }
+      HealthCheckPath : string option
+      IpSecurityRestrictions : IpSecurityRestriction list }
     /// Shorthand for SiteType.ResourceType
     member this.ResourceType = this.SiteType.ResourceType
     /// Shorthand for SiteType.ResourceName
@@ -272,6 +273,16 @@ type Site =
                            javaContainer = this.JavaContainer |> Option.toObj
                            javaContainerVersion = this.JavaContainerVersion |> Option.toObj
                            phpVersion = this.PhpVersion |> Option.toObj
+                           ipSecurityRestrictions = 
+                                match this.IpSecurityRestrictions with
+                                | [] -> null
+                                | restrictions ->
+                                    restrictions
+                                    |> List.mapi (fun index restriction ->
+                                        {| ipAddress = IPAddressCidr.format restriction.IpAddressCidr
+                                           name = restriction.Name
+                                           action = restriction.Action.ToString()
+                                           priority = index + 1 |}) |> box
                            pythonVersion = this.PythonVersion |> Option.toObj
                            http20Enabled = this.HTTP20Enabled |> Option.toNullable
                            webSocketsEnabled = this.WebSocketsEnabled |> Option.toNullable