+++ src/Farmer/Builders/Builders.Functions.fs
@@ -273,7 +273,8 @@ type FunctionsConfig =
                         Some sc
                     | _ -> None
                   WorkerProcess = this.CommonWebConfig.WorkerProcess
-                  HealthCheckPath = this.CommonWebConfig.HealthCheckPath }
+                  HealthCheckPath = this.CommonWebConfig.HealthCheckPath
+                  IpSecurityRestrictions = this.CommonWebConfig.IpSecurityRestrictions }
 
             match this.CommonWebConfig.ServicePlan with
             | DeployableResource this.Name.ResourceName resourceId ->
@@ -343,7 +344,8 @@ type FunctionsBuilder() =
               Slots = Map.empty
               WorkerProcess = None
               ZipDeployPath = None
-              HealthCheckPath = None }
+              HealthCheckPath = None 
+              IpSecurityRestrictions = [] }
           StorageAccount = derived (fun config ->
             let storage = config.Name.ResourceName.Map (sprintf "%sstorage") |> sanitiseStorage |> ResourceName
             storageAccounts.resourceId storage)