+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@ Release Notes
 =============
 ## vNext
 * WebApp: Support custom port for docker container with `docker_port`
+* WebApps/Functions: Add support for ip-restriction rules
 * WebApps/Functions: Don't turn on Logging Extension for Linux App Service.
 
 ## 1.6.26